Organized files to fit Routing design.
